Check available important files before running program

+++ ReversiMPI/run-mpi.sh
@@ -3,21 +3,36 @@
 infoPrompt() {
     case $1 in  
         ("run") 
-            echo "========================"
-            echo "Welcome in Othello!!!"
-            echo "========================"
+            echo "========================="
+            echo "  Welcome in Othello!!!  "
+            echo " "
+            echo "     ● ● ● ● ○ ○ ○ ○     "
+            echo "     ● ● ● ● ○ ○ ○ ○     "
+            echo "     ● ● ● ● ○ ○ ○ ○     "
+            echo "     ● ● ● ● ○ ○ ○ ○     "
+            echo "     ○ ○ ○ ○ ● ● ● ●     " 
+            echo "     ○ ○ ○ ○ ● ● ● ●     " 
+            echo "     ○ ○ ○ ○ ● ● ● ●     " 
+            echo "     ○ ○ ○ ○ ● ● ● ●     "
+            echo " "
+            echo "========================="
         ;; 
         ("end") 
-            echo "Bye bye!"
+            echo "========================="
+            echo "        Bye bye!         "
+            echo "========================="
         ;;
-        (*) echo "Bye bye!" 
+        (*) 
+            echo "========================="
+            echo "        Bye bye!         "
+            echo "========================="
         ;; 
     esac
 }
 
 askPrompt(){
     echo " "
-    read -p "Start a new game? (y/n) " newGame </dev/tty
+    read -p "Start a new game? (y/n): " newGame </dev/tty
     case $newGame in  
         y|Y|j|J) checkHosts ;; 
         n|N) infoPrompt "end" ;;
@@ -26,30 +41,43 @@ askPrompt(){
 }
 
 checkHosts(){
-    #remove hosts file
-    rm ./files
+    # check if file hosts exists
+    if [ ! -f ./hosts ]; then
+        echo "File 'hosts' is not found! Please create one."
+        infoPrompt "end"
+    elif [ ! -s ./hosts ]; then # check if file hosts empty
+        echo "File 'hosts' is empty. Please insert available hosts."
+        infoPrompt "end"
+    else 
+        # if hosts file exists then remove it
+        if [ -e ./files ]; then
+            #remove hosts file
+            rm ./files
+        fi
 
-    #write new hosts file
-    while IFS='' read -r line || [[ -n "$line" ]]; do
-	ping -c1 -W1 $line  && 
-	echo " " && 
-	echo "	CONNECTED TO $line" && 
-	echo -e "$line" >> files || 
-	echo "	CANNOT REACH $line" && 
-	echo " "
-    done < "hosts"
+        #write new hosts file
+        while IFS='' read -r line || [[ -n "$line" ]]; do
+            echo $line
+            ping -c1 -W1 $line  && 
+            echo " " && 
+            echo "	CONNECTED TO $line" && 
+            echo -e "$line" >> files || 
+            echo "	CANNOT REACH $line" && 
+            echo " "
+        done < "hosts"
 
-    sleep 1
+        sleep 1
 
-    echo "AVAILABLE HOSTS: "
-    cat files
-    echo " "
+        echo "AVAILABLE HOSTS: "
+        cat files
+        echo " "
 
-    askNode
+        askNode
+    fi
 }
 
 askNode(){
-    read -p "How many number of processes should we use? " nodesNumber </dev/tty
+    read -p "How many number of processes should we use?: " nodesNumber </dev/tty
     checkGivenNodesNumber
 }
 
@@ -69,7 +97,7 @@ checkGivenNodesNumber(){
 }
 
 askPromptReady(){
-    read -p "Ready? (y/n) " ready </dev/tty
+    read -p "Ready? (y/n): " ready </dev/tty
     case $ready in  
         y|Y|j|J) startProcesses ;; 
         n|N) askPromptReady;;
